Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/script-src-self-blocked-01-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8000/security/contentSecurityPolicy/resources/script.js because it does not appear in the script-src directive of the Content Security Policy.


--------
Frame: '<!--frame1-->'
--------
PASS
